Fix get_borders method
The get_borders method scans through all the layers of guides and
markers to get the max/min of x and y coordinates. The z-coordinates is
fixed by the thickness of the glass so it is not checked.

A flag is also added to the method for saving images in order to toggle
the tight axis options.
Other minor changes to the example code has been done as well.