Added 3D plot and fix 2D plot
Photonic circuits can be plotted and viewed in 3D. The 3D plot shows an
interactive 3D view (where the circuit can be inspected in different
angles using the mouse) and three fixed panes with the projections of
the axis-planes.

2D view shows the xy plane by default. In the future the possibility to
draw markers will be added.
The 2D plot can be saved with the dedicated method. Additional
parameters has been added to allow control over the aspect ratio and the
padding of the frame of the final picture.

Other minor changes has been done and an example function has been added
for rapid tests of the functionality of the class.