Add different waveguide to single layer
Change the get_bounds() method and all the guide's plotting method to
support this feature.

The Cell class has a dictionary of waveguides. Up to now all the
waveguides had to live in different layers of the Cell. This was not ideal
and difficult to manage for complicated circuits or big fabrication (with
lots of circuits).

With this implementation all the waveguides of a circuit can
be added to the same layer indexed by a particular key in the dictionary.
In this way we can add all the waveguides of a circuit to the same
layer and thus using the key to refer to the circuit.

Adding different circuits to the cell or selecting particular circuit
from the cell is very easy with this implementation.